pyproject.toml
[tool.poetry]
name = "devicebay"
version = "0.1.23"
description = "Devices for AI agents"
authors = ["Patrick Barker <patrickbarkerco@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.29"
playwright = {version = "^1.43.0", extras = ["browser"]}
pygithub = {version = "^2.3.0", extras = ["github"]}
toolfuse = "^0.1.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"